feat: add passwordEncoder & AuthenticationManager, Provider
